set login persistence based on platform
